# backend/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AdaptationBatch": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "createdBy": {
            "title": "Createdby",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/patty__adaptation__llm__dummy__DummyModel"
              },
              {
                "$ref": "#/components/schemas/MistralAiModel"
              },
              {
                "$ref": "#/components/schemas/OpenAiModel"
              }
            ],
            "title": "Model"
          },
          "strategySettingsIdentity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Identity"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "createdBy",
          "createdAt",
          "model",
          "strategySettingsIdentity"
        ],
        "title": "AdaptationBatch",
        "type": "object"
      },
...
      "ApiInput": {
        "additionalProperties": false,
        "properties": {
          "exerciseNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exercisenumber"
          },
          "pageNumber": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pagenumber"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "pageNumber",
          "exerciseNumber",
          "text"
        ],
        "title": "ApiInput",
        "type": "object"
      },
...
      "ApiStrategy-Input": {
        "additionalProperties": false,
        "properties": {
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/patty__adaptation__llm__dummy__DummyModel"
              },
              {
                "$ref": "#/components/schemas/MistralAiModel"
              },
              {
                "$ref": "#/components/schemas/OpenAiModel"
              }
            ],
            "title": "Model"
          },
          "settings": {
            "$ref": "#/components/schemas/ApiStrategySettings-Input"
          }
        },
        "required": [
          "model",
          "settings"
        ],
        "title": "ApiStrategy",
        "type": "object"
      },
...
      "BaseAdaptationBatch": {
        "additionalProperties": false,
        "properties": {
          "availableStrategySettings": {
            "items": {
              "$ref": "#/components/schemas/ApiStrategySettings-Output"
            },
            "title": "Availablestrategysettings",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/ApiInput"
            },
            "title": "Inputs",
            "type": "array"
          },
          "strategy": {
            "$ref": "#/components/schemas/ApiStrategy-Output"
          }
        },
        "required": [
          "id",
          "strategy",
          "inputs",
          "availableStrategySettings"
        ],
        "title": "BaseAdaptationBatch",
        "type": "object"
      },
...
      "GetAdaptationBatchResponse": {
        "additionalProperties": false,
        "properties": {
          "createdBy": {
            "title": "Createdby",
            "type": "string"
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/patty__api_router__sandbox_adaptation__GetAdaptationBatchResponse__Exercise"
            },
            "title": "Exercises",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "needsRefresh": {
            "title": "Needsrefresh",
            "type": "boolean"
          },
          "strategy": {
            "$ref": "#/components/schemas/ApiStrategy-Output"
          }
        },
        "required": [
          "id",
          "needsRefresh",
          "createdBy",
          "strategy",
          "exercises"
        ],
        "title": "GetAdaptationBatchResponse",
        "type": "object"
      },
      "GetAdaptationBatchesResponse": {
        "additionalProperties": false,
        "properties": {
          "adaptationBatches": {
            "items": {
              "$ref": "#/components/schemas/AdaptationBatch"
            },
            "title": "Adaptationbatches",
            "type": "array"
          },
          "nextChunkId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextchunkid"
          }
        },
        "required": [
          "adaptationBatches",
          "nextChunkId"
        ],
        "title": "GetAdaptationBatchesResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonSchemaLlmResponseSpecification": {
        "additionalProperties": false,
        "properties": {
          "exampleComponents": {
            "$ref": "#/components/schemas/ExampleComponents"
          },
          "formalism": {
            "const": "json-schema",
            "title": "Formalism",
            "type": "string"
          },
          "format": {
            "const": "json",
            "title": "Format",
            "type": "string"
          },
          "hintComponents": {
            "$ref": "#/components/schemas/HintComponents"
          },
          "instructionComponents": {
            "$ref": "#/components/schemas/InstructionComponents"
          },
          "referenceComponents": {
            "$ref": "#/components/schemas/ReferenceComponents"
          },
          "statementComponents": {
            "$ref": "#/components/schemas/StatementComponents"
          }
        },
        "required": [
          "format",
          "formalism",
          "instructionComponents",
          "exampleComponents",
          "hintComponents",
          "statementComponents",
          "referenceComponents"
        ],
        "title": "JsonSchemaLlmResponseSpecification",
        "type": "object"
      },
...
      "MistralAiModel": {
        "properties": {
          "name": {
            "enum": [
              "mistral-large-2411",
              "mistral-small-2501"
            ],
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "const": "mistralai",
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "name"
        ],
        "title": "MistralAiModel",
        "type": "object"
      },
...
      "OpenAiModel": {
        "properties": {
          "name": {
            "enum": [
              "gpt-4o-2024-08-06",
              "gpt-4o-mini-2024-07-18"
            ],
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "const": "openai",
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "name"
        ],
        "title": "OpenAiModel",
        "type": "object"
      },
...
      "PostAdaptationBatchRequest": {
        "additionalProperties": false,
        "properties": {
          "creator": {
            "title": "Creator",
            "type": "string"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/ApiInput"
            },
            "title": "Inputs",
            "type": "array"
          },
          "strategy": {
            "$ref": "#/components/schemas/ApiStrategy-Input"
          }
        },
        "required": [
          "creator",
          "strategy",
          "inputs"
        ],
        "title": "PostAdaptationBatchRequest",
        "type": "object"
      },
      "PostAdaptationBatchResponse": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "PostAdaptationBatchResponse",
        "type": "object"
      },
...
      "PutAdaptableExerciseClassRequest": {
        "additionalProperties": false,
        "properties": {
          "className": {
            "title": "Classname",
            "type": "string"
          },
          "creator": {
            "title": "Creator",
            "type": "string"
          }
        },
        "required": [
          "creator",
          "className"
        ],
        "title": "PutAdaptableExerciseClassRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "patty__adaptation__llm__dummy__DummyModel": {
        "properties": {
          "name": {
            "enum": [
              "dummy-1",
              "dummy-2",
              "dummy-3",
              "dummy-for-images"
            ],
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "const": "dummy",
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "name"
        ],
        "title": "DummyModel",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/adaptable-exercises/{id}/exercise-class": {
      "put": {
        "operationId": "put_adaptable_exercise_class_api_adaptable_exercises__id__exercise_class_put",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutAdaptableExerciseClassRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Put Adaptable Exercise Class"
      }
    },
    "/api/adaptation-batches": {
      "get": {
        "operationId": "get_adaptation_batches_api_adaptation_batches_get",
        "parameters": [
          {
            "in": "query",
            "name": "chunkId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chunkid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAdaptationBatchesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Adaptation Batches"
      },
      "post": {
        "operationId": "post_adaptation_batch_api_adaptation_batches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostAdaptationBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostAdaptationBatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Post Adaptation Batch"
      }
    },
    "/api/adaptation-batches/{id}": {
      "get": {
        "operationId": "get_adaptation_batch_api_adaptation_batches__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAdaptationBatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Adaptation Batch"
      }
    },
    "/api/adaptation-llm-response-schema": {
      "post": {
        "operationId": "make_adaptation_llm_response_schema_api_adaptation_llm_response_schema_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonSchemaLlmResponseSpecification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Make Adaptation Llm Response Schema Api Adaptation Llm Response Schema Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Make Adaptation Llm Response Schema"
      }
    },
...
    "/api/available-adaptation-llm-models": {
      "get": {
        "operationId": "get_available_adaptation_llm_models_api_available_adaptation_llm_models_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/patty__adaptation__llm__dummy__DummyModel"
                      },
                      {
                        "$ref": "#/components/schemas/MistralAiModel"
                      },
                      {
                        "$ref": "#/components/schemas/OpenAiModel"
                      }
                    ]
                  },
                  "title": "Response Get Available Adaptation Llm Models Api Available Adaptation Llm Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Available Adaptation Llm Models"
      }
    },
...
    "/api/base-adaptation-batch": {
      "get": {
        "operationId": "get_base_adaptation_batch_api_base_adaptation_batch_get",
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "required": true,
            "schema": {
              "title": "User",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "base",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Base"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAdaptationBatch"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Base Adaptation Batch"
      }
    },
...
    "/api/exercise-classes": {
      "get": {
        "operationId": "get_exercise_classes_api_exercise_classes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Exercise Classes Api Exercise Classes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Exercise Classes"
      }
    },
...
  }
}